% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality.R
\name{data_quality}
\alias{data_quality}
\title{Data quality report}
\usage{
data_quality(data)
}
\arguments{
\item{data}{A dataframe}
}
\value{
A list containing data quality information
}
\description{
This function gives a quick overview of the quality of a dataframe,
including number of rows, columns, missing values, and duplicate rows.
}
