% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{datacleaner_server}
\alias{datacleaner_server}
\title{Server logic for datacleaner gadget}
\usage{
datacleaner_server(input, output, session)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}
}
\description{
Server logic for datacleaner gadget
}
