% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_column_info}
\alias{get_column_info}
\title{Get column information}
\usage{
get_column_info(data)
}
\arguments{
\item{data}{Data frame}
}
\value{
Data frame with column metadata
}
\description{
Get column information
}
