% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{import_csv}
\alias{import_csv}
\title{Import CSV with encoding detection}
\usage{
import_csv(file_path)
}
\arguments{
\item{file_path}{Path to CSV file}
}
\value{
A data frame
}
\description{
Import CSV with encoding detection
}
