% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{import_data}
\alias{import_data}
\title{Import data from file}
\usage{
import_data(file_path)
}
\arguments{
\item{file_path}{Path to the data file}
}
\value{
A data frame
}
\description{
Import data from file
}
