% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{insert_code_to_rstudio}
\alias{insert_code_to_rstudio}
\title{Insert code into RStudio editor}
\usage{
insert_code_to_rstudio(code)
}
\arguments{
\item{code}{Character string with R code}
}
\value{
Invisible NULL
}
\description{
Insert code into RStudio editor
}
