% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanimpute.R
\name{meanimpute}
\alias{meanimpute}
\title{Meanimputation}
\usage{
meanimpute(x)
}
\arguments{
\item{x}{A numeric vector containing real numbers}
}
\value{
Imputed vector \code{x} where NAs are replaced with the mean of \code{x}
}
\description{
Replaces NA values with the mean of \code{x}
}
\examples{
meanimpute(x = c(1, NA, 3, 4, 5, 6, 7, 8, NA, 10))
meanimpute(c(1, NA, 3, 4, 5, 6, 7, 8, NA, 10))
}
