% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicates.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{Remove duplicate rows}
\usage{
remove_duplicates(data)
}
\arguments{
\item{data}{A dataframe}
}
\value{
A cleaned dataframe without duplicate rows
}
\description{
This function removes duplicate rows from a dataframe.
}
