% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_empty.R
\name{remove_empty}
\alias{remove_empty}
\title{Remove columns with only missing values}
\usage{
remove_empty(data)
}
\arguments{
\item{data}{A dataframe}
}
\value{
A dataframe without empty columns
}
\description{
This function removes columns that contain only NA values.
}
