% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_na.R
\name{remove_na}
\alias{remove_na}
\title{Remove rows with missing values}
\usage{
remove_na(data)
}
\arguments{
\item{data}{A dataframe}
}
\value{
A cleaned dataframe without missing values
}
\description{
This function removes rows containing missing values from a dataframe.
}
