% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{settingsServer}
\alias{settingsServer}
\title{Settings Server Module}
\usage{
settingsServer(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
Module server function
}
\description{
Settings Server Module
}
