% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_log.R
\name{transform_log}
\alias{transform_log}
\title{Transform_log}
\usage{
transform_log(x)
}
\arguments{
\item{x}{A numeric vector containing positive real numbers}
}
\value{
Natural logarithm of the values in \code{x}.
}
\description{
Transforms values to natural logarithm.
}
\examples{
transform_log(x = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10))
transform_log(c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10))
}
