% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windsorize.R
\name{windsorize}
\alias{windsorize}
\title{Windsorize}
\usage{
windsorize(x, p = 0.9)
}
\arguments{
\item{x}{A vector.}

\item{p}{A quantile.}
}
\value{
inuput vector x with trimmed outliers by (1-p) percentile.
}
\description{
Transform all outliner data to
(1-p)/2 percentile value for lower outliers and 
(1+p)/2 for higher outliers.
}
\examples{
windsorize(rnorm(100,0,1))
}
