% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanimpute.R
\name{meanimpute}
\alias{meanimpute}
\title{Meanimputation}
\usage{
meanimpute(x)
}
\arguments{
\item{x}{A numeric vector containing the data}
}
\value{
the input vector with its NA values replaced by the mean
}
\description{
Meanimputation
}
\examples{
log(1:5)
}
