% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanimpute.R
\name{meanimpute}
\alias{meanimpute}
\title{Meanimpute}
\usage{
meanimpute(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
new vector, where NA's are replace by mean \code{x}
}
\description{
Replace NA's with mean value
}
\examples{
example_vector=c(1,5,NA,NA)
meanimpute(example_vector)
}
