% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_log.R
\name{transform_log}
\alias{transform_log}
\title{Log Transform
 
Transform numerical values into their log values.}
\usage{
transform_log(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The log values of \code{x}.
}
\description{
Log Transform
 
Transform numerical values into their log values.
}
\examples{
transform_log(exp(rnorm(7)))
}
