% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_log.R
\name{transform_log}
\alias{transform_log}
\title{transform_log 
Transform numerical values into their log values}
\usage{
transform_log(x)
}
\arguments{
\item{x}{A vector}
}
\value{
logarithm of x
}
\description{
transform_log 
Transform numerical values into their log values
}
\examples{
transform_log(c(NA,0,-1,exp(2)))
}
