% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_log.R
\name{transform_log}
\alias{transform_log}
\title{transform_log}
\usage{
transform_log(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
log-transformed vector \code{x}
}
\description{
log-transformation of a numeric vector. For details about log-transformation please see you basic school math textbook.
}
\examples{
example_vector=c(1,2,3,4,5,6,7,8,9,10)
transform_log(example_vector)
}
