% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_log.R
\name{transform_log}
\alias{transform_log}
\title{Log transforms data.}
\usage{
transform_log(x)
}
\arguments{
\item{x}{A numeric vector containing the data}
}
\value{
A numeric vector whose entries are the logarithms of the input.
}
\description{
Log transforms data.
}
\examples{
log(1:5)
}
