% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windsorize.R
\name{windsorize}
\alias{windsorize}
\title{Windsorize}
\usage{
windsorize(x, p = 0.9)
}
\arguments{
\item{x}{A vector.}

\item{p}{A quantile.}
}
\description{
Winsorizing or winsorization is the transformation of statistics by limiting
extreme values in the statistical data to reduce the effect of possibly
spurious outliers.
}
\examples{
windsorize(rnorm(5))

}
