% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windsorize.R
\name{windsorize}
\alias{windsorize}
\title{Windsorize}
\usage{
windsorize(x, p = 0.9)
}
\arguments{
\item{x}{A vector.}

\item{p}{A quantile.}
}
\value{
dataset with trimmed outliers with 10% percentile
}
\description{
set all outliers to a specified percentile of the data; 
a 90% winsorization would see all data below the 5th percentile set 
to the 5th percentile, 
and data above the 95th percentile set to the 95th percentile.
}
\examples{
windsorize(c(92, 19, 101, 58, 101, 91, 26, 78, 10, 13, −5, 101, 86, −5))
}
