% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windsorize.R
\name{windsorize}
\alias{windsorize}
\title{Windsorize}
\usage{
windsorize(x, p = 0.9)
}
\arguments{
\item{x}{A numerical vector.}

\item{p}{Quantile value for outliers removal.}
}
\description{
Replacing values of vector \code{x} greater or smaller then \code{q} quantile values.
}
\examples{
 
windsorize(c(1,499,500,501))
}
